# docs/incident-ai-video-investigation-openapi.yaml
openapi: 3.1.0
info:
  title: Incident AI Video Investigation API
  version: 1.0.0
  description: |
    Backend contract for AI-assisted incident evidence analysis using uploaded photos and videos.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
tags:
  - name: Incident AI
paths:
  /incidents/{incidentId}/ai-evidence/upload-session:
    post:
      tags: [Incident AI]
      summary: Create signed upload targets for incident evidence
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUploadSessionRequest'
      responses:
        '200':
          description: Upload session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUploadSessionResponse'
  /incidents/{incidentId}/ai-evidence/confirm:
    post:
      tags: [Incident AI]
      summary: Confirm uploaded incident evidence metadata
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmEvidenceRequest'
      responses:
        '200':
          description: Evidence confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmEvidenceResponse'
  /incidents/{incidentId}/ai-analysis:
    post:
      tags: [Incident AI]
      summary: Start AI investigation analysis job
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAnalysisRequest'
      responses:
        '202':
          description: Analysis queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartAnalysisResponse'
    get:
      tags: [Incident AI]
      summary: Get completed analysis result
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResultResponse'
  /incidents/{incidentId}/ai-analysis/status:
    get:
      tags: [Incident AI]
      summary: Get analysis job status
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Analysis status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisStatusResponse'
  /incidents/{incidentId}/ai-analysis/retry:
    post:
      tags: [Incident AI]
      summary: Retry AI analysis
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetryAnalysisRequest'
      responses:
        '202':
          description: Retry queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartAnalysisResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IncidentId:
      name: incidentId
      in: path
      required: true
      schema:
        type: string
  schemas:
    UploadTarget:
      type: object
      properties:
        storagePath:
          type: string
        uploadUrl:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
    FileDescriptor:
      type: object
      properties:
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
      required: [fileName, mimeType, sizeBytes]
    ConfirmedEvidenceFile:
      type: object
      properties:
        storagePath:
          type: string
        fileName:
          type: string
        mimeType:
          type: string
      required: [storagePath, fileName, mimeType]
    CreateUploadSessionRequest:
      type: object
      properties:
        photo:
          $ref: '#/components/schemas/FileDescriptor'
        video:
          $ref: '#/components/schemas/FileDescriptor'
      required: [photo]
    CreateUploadSessionResponse:
      type: object
      properties:
        incidentId:
          type: string
        uploadSessionId:
          type: string
        photo:
          $ref: '#/components/schemas/UploadTarget'
        video:
          $ref: '#/components/schemas/UploadTarget'
        expiresAt:
          type: string
          format: date-time
      required: [incidentId, uploadSessionId, photo, expiresAt]
    ConfirmEvidenceRequest:
      type: object
      properties:
        uploadSessionId:
          type: string
        photo:
          $ref: '#/components/schemas/ConfirmedEvidenceFile'
        video:
          $ref: '#/components/schemas/ConfirmedEvidenceFile'
        notes:
          type: string
      required: [uploadSessionId, photo]
    ConfirmEvidenceResponse:
      type: object
      properties:
        incidentId:
          type: string
        evidenceStatus:
          type: string
          enum: [confirmed]
        photoAttached:
          type: boolean
        videoAttached:
          type: boolean
      required: [incidentId, evidenceStatus, photoAttached, videoAttached]
    IncidentContext:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        equipmentInvolved:
          type: string
        immediateAction:
          type: string
        smartCategory:
          type: string
        severity:
          type: string
        type:
          type: string
    StartAnalysisRequest:
      type: object
      properties:
        forceRerun:
          type: boolean
        includeVideo:
          type: boolean
        includeAudioTranscript:
          type: boolean
        frameSampleSeconds:
          type: integer
        maxFrames:
          type: integer
        analysisLanguage:
          type: string
        incidentContext:
          $ref: '#/components/schemas/IncidentContext'
    StartAnalysisResponse:
      type: object
      properties:
        incidentId:
          type: string
        jobId:
          type: string
        status:
          type: string
          enum: [queued, processing]
      required: [incidentId, jobId, status]
    AnalysisStatusResponse:
      type: object
      properties:
        incidentId:
          type: string
        jobId:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed, cancelled]
        stage:
          type: string
        progressPercent:
          type: integer
        startedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [incidentId, jobId, status]
    TranscriptSegment:
      type: object
      properties:
        startMs:
          type: integer
        endMs:
          type: integer
        speaker:
          type: string
        text:
          type: string
    TranscriptPayload:
      type: object
      properties:
        text:
          type: string
        segments:
          type: array
          items:
            $ref: '#/components/schemas/TranscriptSegment'
    DraftCapa:
      type: object
      properties:
        act:
          type: string
        priority:
          type: string
    FishbonePayload:
      type: object
      properties:
        man:
          type: array
          items: { type: string }
        machine:
          type: array
          items: { type: string }
        material:
          type: array
          items: { type: string }
        method:
          type: array
          items: { type: string }
        environment:
          type: array
          items: { type: string }
    InvestigationDraft:
      type: object
      properties:
        eventSummary:
          type: string
        visibleHazards:
          type: array
          items: { type: string }
        equipmentCondition:
          type: array
          items: { type: string }
        immediateCauses:
          type: array
          items: { type: string }
        contributingFactors:
          type: array
          items: { type: string }
        fiveWhys:
          type: array
          items: { type: string }
        fishbone:
          $ref: '#/components/schemas/FishbonePayload'
        rootCause:
          type: string
        capa:
          type: array
          items:
            $ref: '#/components/schemas/DraftCapa'
        confidence:
          type: string
        missingInformation:
          type: array
          items: { type: string }
    AnalysisResultResponse:
      type: object
      properties:
        incidentId:
          type: string
        status:
          type: string
          enum: [completed]
        provider:
          type: string
        transcriptionModel:
          type: string
        visionModel:
          type: string
        transcript:
          $ref: '#/components/schemas/TranscriptPayload'
        draft:
          $ref: '#/components/schemas/InvestigationDraft'
        review:
          type: object
          properties:
            status:
              type: string
              enum: [pending, accepted, edited, rejected]
    RetryAnalysisRequest:
      type: object
      properties:
        reason:
          type: string
        override:
          type: object
          properties:
            frameSampleSeconds:
              type: integer
            maxFrames:
              type: integer